Move fixtures commands to a script
