+++ docker-compose.yml
@@ -86,7 +86,6 @@ services:
     command: >-
       bash -c "
       python manage.py migrate
-      && python manage.py init_gc_demo
       && python manage.py runserver_plus 0.0.0.0:8000 --nopin
       "
     restart: always
